requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dateutil>=2.8.0
loguru>=0.7.0
python-dotenv>=1.0.0
httpx>=0.25.0
aiohttp>=3.9.0
apscheduler>=3.10.0
langchain-core>=0.1.0

# 认证相关
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.6

# 情感分析相关
jieba>=0.42.1
snownlp>=0.12.3
